server/requirements.txt
# Core web framework
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.6

# Configuration
python-dotenv==1.0.1

# Data validation
pydantic==2.5.0

# OpenAI and LLM
openai==1.3.7
tenacity==8.5.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Caching and task queue
redis==5.0.4
celery==5.3.6

# HTTP client
httpx>=0.24.0,<0.25.0

# ML and embeddings (optional - for advanced features)
sentence-transformers==2.2.2
numpy==1.26.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
jinja2==3.1.2
markdown==3.5.1
pyyaml==6.0.1

# Error monitoring
sentry-sdk[fastapi]==2.0.0

# External services (optional)
supabase==2.0.2
stripe==7.8.0
sendgrid==6.10.0
boto3==1.34.0
docker==6.1.3
